Create :core module, allocate common functionality to :core module
