+++ apod/src/main/java/com/project/apod/viewmodel/APODViewModelFactory.kt
@@ -2,6 +2,7 @@ package com.project.apod.viewmodel
 
 import androidx.lifecycle.SavedStateHandle
 import com.project.apod.domain.Repository
+import com.project.core.viewmodel.ViewModelAssistedFactory
 
 class APODViewModelFactory(
     private val repository: Repository
