+++ core/src/main/java/com/project/core/viewmodel/ViewModelAssistedFactory.kt
@@ -1,4 +1,4 @@
-package com.project.apod.viewmodel
+package com.project.core.viewmodel
 
 import androidx.lifecycle.SavedStateHandle
